[BACKLOG-19621] updates step archetype plugin structure.
