+++ pentaho-pdi-job-plugin-archetype/src/main/resources/META-INF/maven/archetype-metadata.xml
@@ -1,51 +1,83 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<archetype-descriptor xsi:schemaLocation="http://maven.apache.org/plugins/maven-archetype-plugin/archetype-descriptor/1.0.0 http://maven.apache.org/xsd/archetype-descriptor-1.0.0.xsd" name="pdi-job-plugin"
-    xmlns="http://maven.apache.org/plugins/maven-archetype-plugin/archetype-descriptor/1.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+<archetype-descriptor
+  xsi:schemaLocation="http://maven.apache.org/plugins/maven-archetype-plugin/archetype-descriptor/1.0.0 http://maven.apache.org/xsd/archetype-descriptor-1.0.0.xsd"
+  name="pdi-job-plugin"
+  xmlns="http://maven.apache.org/plugins/maven-archetype-plugin/archetype-descriptor/1.0.0"
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
   <requiredProperties>
-    <requiredProperty key="plugin_class_name"></requiredProperty>
-    <requiredProperty key="plugin_name"></requiredProperty>
-    <requiredProperty key="plugin_category"></requiredProperty>
-    <requiredProperty key="plugin_description"></requiredProperty>
+    <requiredProperty key="plugin_class_name"/>
+    <requiredProperty key="plugin_name"/>
+    <requiredProperty key="plugin_category"/>
+    <requiredProperty key="plugin_description"/>
     <requiredProperty key="kettleVersion">
-      <defaultValue>7.0.0.3-62</defaultValue>
-    </requiredProperty>    
+      <defaultValue>8.1-SNAPSHOT</defaultValue>
+    </requiredProperty>
+
+    <requiredProperty key="package">
+      <defaultValue>org.pentaho.di</defaultValue>
+    </requiredProperty>
   </requiredProperties>
+
   <fileSets>
-    <fileSet filtered="true" packaged="true" encoding="UTF-8">
-      <directory>src/main/java</directory>
-      <includes>
-        <include>**/*.java</include>
-      </includes>
-    </fileSet>
-    <fileSet filtered="true" packaged="true" encoding="UTF-8">
-      <directory>src/test/java</directory>
-      <includes>
-        <include>**/*.java</include>
-      </includes>
-    </fileSet>
-    <fileSet filtered="true" encoding="UTF-8">
-      <directory>src/main/resources</directory>
-      <includes>
-        <include>**/*.xml</include>
-        <include>**/*.svg</include>
-      </includes>
-    </fileSet>
-    <fileSet filtered="true" packaged="true" encoding="UTF-8">
-      <directory>src/main/resources</directory>
-      <includes>
-        <include>**/messages_*.properties</include>
-      </includes>
-    </fileSet>
     <fileSet encoding="UTF-8">
-      <directory></directory>
+      <directory/>
       <includes>
         <include>.gitignore</include>
         <include>README.md</include>
+        <include>LICENSE.txt</include>
       </includes>
-      <excludes>
-        <exclude>*.iml</exclude>
-      </excludes>
     </fileSet>
   </fileSets>
+
+  <modules>
+    <module id="${rootArtifactId}-impl" dir="impl" name="${plugin_name} Impl">
+      <fileSets>
+        <fileSet filtered="true" packaged="true" encoding="UTF-8">
+          <directory>src/main/java</directory>
+          <includes>
+            <include>**/*.java</include>
+          </includes>
+        </fileSet>
+        <fileSet filtered="true" encoding="UTF-8">
+          <directory>src/main/resources</directory>
+          <includes>
+            <include>**/*.xml</include>
+            <include>**/*.properties</include>
+          </includes>
+        </fileSet>
+      </fileSets>
+    </module>
+    <module id="${rootArtifactId}-ui" dir="ui" name="${plugin_name} UI">
+      <fileSets>
+        <fileSet filtered="true" packaged="true" encoding="UTF-8">
+          <directory>src/main/java</directory>
+          <includes>
+            <include>**/*.java</include>
+          </includes>
+        </fileSet>
+        <fileSet filtered="true" encoding="UTF-8">
+          <directory>src/main/resources</directory>
+          <includes>
+            <include>**/*.properties</include>
+          </includes>
+        </fileSet>
+        <fileSet encoding="UTF-8">
+          <directory>src/main/resources</directory>
+          <includes>
+            <include>**/*.svg</include>
+          </includes>
+        </fileSet>
+      </fileSets>
+    </module>
+    <module id="${rootArtifactId}-plugin" dir="assembly" name="${plugin_name} Assembly">
+      <fileSets>
+        <fileSet filtered="true" encoding="UTF-8">
+          <directory>src/main/feature</directory>
+          <includes>
+            <include>**/*.xml</include>
+          </includes>
+        </fileSet>
+      </fileSets>
+    </module>
+  </modules>
 </archetype-descriptor>
