+++ pentaho-pdi-job-plugin-archetype/src/main/resources/archetype-resources/impl/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -1,0 +1,23 @@
+#set( $symbol_pound = '#' )
+#set( $symbol_dollar = '$' )
+#set( $symbol_escape = '\' )
+<?xml version="1.0" encoding="UTF-8"?>
+<blueprint xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
+           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+           xmlns:pen="http://www.pentaho.com/xml/schemas/pentaho-blueprint"
+           xsi:schemaLocation="
+             http://www.osgi.org/xmlns/blueprint/v1.0.0 http://www.osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd
+             http://www.pentaho.com/xml/schemas/pentaho-blueprint http://www.pentaho.com/xml/schemas/pentaho-blueprint.xsd">
+
+  <bean id="${plugin_class_name}Entry" class="${package}.job.entries.${rootArtifactId}.${plugin_class_name}Entry" scope="prototype">
+    <pen:di-plugin type="org.pentaho.di.core.plugins.JobEntryPluginType"/>
+  </bean>
+
+  <!-- Define a bean for our new entry analyzer -->
+  <bean id="${plugin_class_name}EntryAnalyzer" class="${package}.job.entries.${rootArtifactId}.${plugin_class_name}EntryAnalyzer"/>
+
+  <service id="${plugin_class_name}EntryAnalyzerService"
+           interface="org.pentaho.metaverse.api.analyzer.kettle.jobentry.IJobEntryAnalyzer"
+           ref="${plugin_class_name}EntryAnalyzer"/>
+
+</blueprint>
