+++ pentaho-pdi-job-plugin-archetype/src/main/resources/archetype-resources/impl/src/main/resources/__packageInPathFormat__/job/entries/__rootArtifactId__/messages/messages_en_US.properties
@@ -1,0 +1,22 @@
+#set( $symbol_pound = '#' )
+#set( $symbol_dollar = '$' )
+#set( $symbol_escape = '\' )
+#####################################################################
+##
+##  ${plugin_class_name}
+##
+#####################################################################
+${plugin_class_name}.Name=${plugin_name} 
+${plugin_class_name}.Description=${plugin_description} 
+${plugin_class_name}.TooltipDesc=${plugin_description} 
+${plugin_class_name}.Log.LineNumber=Linenr
+
+#####################################################################
+##
+##  ${plugin_class_name}Entry
+##
+#####################################################################
+${plugin_class_name}Entry.CheckResult.NotReceivingFields=Not receiving any fields from previous jobs\!
+${plugin_class_name}Entry.CheckResult.JobRecevingData=Job is connected to previous one, receiving {0} fields
+${plugin_class_name}Entry.CheckResult.JobRecevingData2=Job is receiving info from other jobs.
+${plugin_class_name}Entry.CheckResult.NoInputReceivedFromOtherJobs=No input received from other jobs\!
