--- pentaho-pdi-job-plugin-archetype/src/main/resources/archetype-resources/src/main/java/Messages.java
@@ -1,56 +1,0 @@
-#set( $symbol_pound = '#' )
-#set( $symbol_dollar = '$' )
-#set( $symbol_escape = '\' )
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package ${package};
-
-import org.pentaho.di.i18n.BaseMessages;
-
-public class Messages {
-  public static final Class<Messages> PKG = Messages.class;
-
-  public static String getString( String key ) {
-    return BaseMessages.getString( PKG, key );
-  }
-
-  public static String getString( String key, String param1 ) {
-    return BaseMessages.getString( PKG, key, param1 );
-  }
-
-  public static String getString( String key, String param1, String param2 ) {
-    return BaseMessages.getString( PKG, key, param1, param2 );
-  }
-
-  public static String getString( String key, String param1, String param2, String param3 ) {
-    return BaseMessages.getString( PKG, key, param1, param2, param3 );
-  }
-
-  public static String getString( String key, String param1, String param2, String param3, String param4 ) {
-    return BaseMessages.getString( PKG, key, param1, param2, param3, param4 );
-  }
-
-  public static String getString( String key, String param1, String param2, String param3, String param4, String param5 ) {
-    return BaseMessages.getString( PKG, key, param1, param2, param3, param4, param5 );
-  }
-
-  public static String getString( String key, String param1, String param2, String param3, String param4, String param5, String param6 ) {
-    return BaseMessages.getString( PKG, key, param1, param2, param3, param4, param5, param6 );
-  }
-}
