--- pentaho-pdi-job-plugin-archetype/src/main/resources/archetype-resources/src/main/java/__plugin_class_name__EntryAnalyzer.java
@@ -1,58 +1,0 @@
-#set( $symbol_pound = '#' )
-#set( $symbol_dollar = '$' )
-#set( $symbol_escape = '\' )
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package ${package};
-
-import org.pentaho.di.core.exception.KettleException;
-import org.pentaho.di.core.exception.KettleMissingPluginsException;
-import org.pentaho.di.core.exception.KettleXMLException;
-import org.pentaho.di.job.entry.JobEntryInterface;
-import org.pentaho.di.repository.Repository;
-import org.pentaho.di.repository.RepositoryDirectoryInterface;
-import org.pentaho.dictionary.DictionaryConst;
-import org.pentaho.metaverse.api.IComponentDescriptor;
-import org.pentaho.metaverse.api.IMetaverseNode;
-import org.pentaho.metaverse.api.MetaverseAnalyzerException;
-import org.pentaho.metaverse.api.MetaverseComponentDescriptor;
-import org.pentaho.metaverse.api.analyzer.kettle.KettleAnalyzerUtil;
-import org.pentaho.metaverse.api.analyzer.kettle.jobentry.JobEntryAnalyzer;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
-import java.util.HashSet;
-import java.util.Set;
-
-public class ${plugin_class_name}EntryAnalyzer extends JobEntryAnalyzer<${plugin_class_name}Entry> {
-
-  private Logger log = LoggerFactory.getLogger( ${plugin_class_name}EntryAnalyzer.class );
-
-  @Override
-  public Set<Class<? extends JobEntryInterface>> getSupportedEntries() {
-    Set<Class<? extends JobEntryInterface>> supportedEntries = new HashSet<Class<? extends JobEntryInterface>>();
-    supportedEntries.add( ${plugin_class_name}Entry.class );
-    return supportedEntries;
-  }
-
-  @Override
-  protected void customAnalyze( ${plugin_class_name}Entry entry, IMetaverseNode rootNode ) throws MetaverseAnalyzerException {
-  }
-}
