--- pentaho-pdi-job-plugin-archetype/src/main/resources/archetype-resources/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -1,39 +1,0 @@
-#set( $symbol_pound = '#' )
-#set( $symbol_dollar = '$' )
-#set( $symbol_escape = '\' )
-<?xml version="1.0" encoding="UTF-8"?>
-<blueprint xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
-           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-           xmlns:pen="http://www.pentaho.com/xml/schemas/pentaho-blueprint"
-           xsi:schemaLocation="
-             http://www.osgi.org/xmlns/blueprint/v1.0.0 http://www.osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd
-             http://www.pentaho.com/xml/schemas/pentaho-blueprint http://www.pentaho.com/xml/schemas/pentaho-blueprint.xsd">
-
-    <bean id="${plugin_class_name}Entry" class="${package}.${plugin_class_name}Entry" scope="prototype"/>
-    <bean id="${plugin_class_name}Plugin" class="org.pentaho.di.osgi.OSGIPlugin" scope="prototype">
-        <property name="mainType" value="org.pentaho.di.job.entry.JobEntryInterface"/>
-        <property name="name" value="${plugin_name}"/>
-        <property name="ID" value="${plugin_class_name}"/>
-        <property name="imageFile" value="${plugin_class_name}.svg"/>
-        <property name="description" value="${plugin_description}"/>
-        <property name="pluginTypeInterface" value="org.pentaho.di.core.plugins.JobEntryPluginType"/>
-        <property name="category" value="${plugin_category}"/>
-        <property name="classToBeanMap">
-            <map>
-                <entry key="org.pentaho.di.job.entry.JobEntryInterface" value="${plugin_class_name}Entry"/>
-            </map>
-        </property>
-    </bean>
-    <service id="${plugin_class_name}PluginService" interface="org.pentaho.di.core.plugins.PluginInterface" ref="${plugin_class_name}Plugin">
-        <service-properties>
-            <entry key="PluginType" value="org.pentaho.di.core.plugins.JobEntryPluginType"/>
-        </service-properties>
-    </service>
-
-    <bean id="${plugin_class_name}EntryAnalyzer" class="${package}.${plugin_class_name}EntryAnalyzer" />
-
-    <service id="${plugin_class_name}EntryAnalyzerService"
-             interface="org.pentaho.metaverse.api.analyzer.kettle.jobentry.IJobEntryAnalyzer"
-             ref="${plugin_class_name}EntryAnalyzer"/>
-
-</blueprint>
