--- pentaho-pdi-job-plugin-archetype/src/main/resources/archetype-resources/src/main/resources/__artifactId__-feature.xml
@@ -1,10 +1,0 @@
-#set( $symbol_pound = '#' )
-#set( $symbol_dollar = '$' )
-#set( $symbol_escape = '\' )
-<features xmlns="http://karaf.apache.org/xmlns/features/v1.0.0">
-
-  <feature name="${artifactId}-feature" version="${version}">
-    <bundle>mvn:${groupId}/${artifactId}/${version}</bundle>
-  </feature>
-
-</features>
