--- pentaho-pdi-job-plugin-archetype/src/main/resources/archetype-resources/src/main/resources/messages/messages_en_US.properties
@@ -1,43 +1,0 @@
-#set( $symbol_pound = '#' )
-#set( $symbol_dollar = '$' )
-#set( $symbol_escape = '\' )
-#####################################################################
-##
-##  ${plugin_class_name}Dialog
-##
-#####################################################################
-${plugin_class_name}EntryDialog.Jobname.Label=Entry name\:
-${plugin_class_name}EntryDialog.Shell.Title=My Job
-${plugin_class_name}EntryDialog.ActionButton=Action
-${plugin_class_name}EntryDialog.GroupText=Fieldset
-${plugin_class_name}EntryDialog.TextField350=Textfield (350px)\:
-${plugin_class_name}EntryDialog.TextField250=Textfield (250px)\:
-${plugin_class_name}EntryDialog.TextField75=Textfield (75px)\:
-${plugin_class_name}EntryDialog.Dropdown=Dropdown\:
-${plugin_class_name}EntryDialog.Tab1=Tab 1
-${plugin_class_name}EntryDialog.Tab2=Tab 2
-${plugin_class_name}EntryDialog.Button=Button
-${plugin_class_name}EntryDialog.RadioButton1=Radio button 1
-${plugin_class_name}EntryDialog.RadioButton2=Radio button 2
-${plugin_class_name}EntryDialog.Checkbox1=Checkbox 1
-${plugin_class_name}EntryDialog.Checkbox2=Checkbox 2
-${plugin_class_name}EntryDialog.TableHeader=Column
-${plugin_class_name}EntryDialog.ComboBoxOption1=Option 1
-${plugin_class_name}EntryDialog.ComboBoxOption2=Option 2
-
-#####################################################################
-##
-##  ${plugin_class_name}
-##
-#####################################################################
-${plugin_class_name}.Log.LineNumber=Linenr
-
-#####################################################################
-##
-##  ${plugin_class_name}Entry
-##
-#####################################################################
-${plugin_class_name}Entry.CheckResult.NotReceivingFields=Not receiving any fields from previous jobs\!
-${plugin_class_name}Entry.CheckResult.JobRecevingData=Job is connected to previous one, receiving {0} fields
-${plugin_class_name}Entry.CheckResult.JobRecevingData2=Job is receiving info from other jobs.
-${plugin_class_name}Entry.CheckResult.NoInputReceivedFromOtherJobs=No input received from other jobs\!
