+++ pentaho-pdi-step-plugin-archetype/src/main/resources/archetype-resources/impl/src/main/java/trans/steps/__rootArtifactId__/__plugin_class_name__Meta.java
@@ -1,0 +1,131 @@
+#set( $symbol_pound = '#' )
+#set( $symbol_dollar = '$' )
+#set( $symbol_escape = '\' )
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package ${package}.trans.steps.${rootArtifactId};
+
+import org.pentaho.di.core.annotations.Step;
+import org.pentaho.di.core.CheckResult;
+import org.pentaho.di.core.CheckResultInterface;
+import org.pentaho.di.core.database.DatabaseMeta;
+import org.pentaho.di.core.exception.KettleException;
+import org.pentaho.di.core.exception.KettleStepException;
+import org.pentaho.di.core.exception.KettleXMLException;
+import org.pentaho.di.core.row.RowMetaInterface;
+import org.pentaho.di.core.row.ValueMetaInterface;
+import org.pentaho.di.core.row.value.ValueMetaFactory;
+import org.pentaho.di.core.variables.VariableSpace;
+import org.pentaho.di.core.xml.XMLHandler;
+import org.pentaho.di.i18n.BaseMessages;
+import org.pentaho.di.repository.ObjectId;
+import org.pentaho.di.repository.Repository;
+import org.pentaho.di.trans.Trans;
+import org.pentaho.di.trans.TransMeta;
+import org.pentaho.di.trans.step.BaseStepMeta;
+import org.pentaho.di.trans.step.StepInterface;
+import org.pentaho.di.trans.step.StepMeta;
+import org.pentaho.di.trans.step.StepMetaInterface;
+import org.pentaho.di.trans.step.StepDataInterface;
+import org.pentaho.metastore.api.IMetaStore;
+import org.w3c.dom.Node;
+
+import java.util.ArrayList;
+import java.util.List;
+
+
+/**
+ * Skeleton for PDI Step plugin.
+ */
+@Step( id = "${plugin_class_name}", image = "${rootArtifactId}/ui/images/${plugin_class_name}.svg", i18nPackageName ="${package}.trans.steps.${rootArtifactId}",
+  name = "${plugin_class_name}.Name", description = "${plugin_class_name}.Description",
+  categoryDescription = "i18n:org.pentaho.di.trans.step:BaseStep.Category.${plugin_category}" )
+public class ${plugin_class_name}Meta extends BaseStepMeta implements StepMetaInterface {
+  
+  private static Class<?> PKG = ${plugin_class_name}.class; // for i18n purposes, needed by Translator2!!
+
+  public ${plugin_class_name}Meta() {
+    super(); // allocate BaseStepMeta
+  }
+
+  public void loadXML( Node stepnode, List<DatabaseMeta> databases, IMetaStore metaStore ) throws KettleXMLException {
+    readData( stepnode );
+  }
+
+  public Object clone() {
+    Object retval = super.clone();
+    return retval;
+  }
+  
+  private void readData( Node stepnode ) {
+    // Parse the XML (starting with the given stepnode) to extract the step metadata (into member variables, for example)
+  }
+
+  public void setDefault() {
+  }
+
+  public void readRep( Repository rep, IMetaStore metaStore, ObjectId id_step, List<DatabaseMeta> databases ) throws KettleException {
+  }
+  
+  public void saveRep( Repository rep, IMetaStore metaStore, ObjectId id_transformation, ObjectId id_step )
+    throws KettleException {
+  }
+  
+  public void getFields( RowMetaInterface rowMeta, String origin, RowMetaInterface[] info, StepMeta nextStep, 
+    VariableSpace space, Repository repository, IMetaStore metaStore ) throws KettleStepException {
+    // Default: nothing changes to rowMeta
+  }
+  
+  public void check( List<CheckResultInterface> remarks, TransMeta transMeta, 
+    StepMeta stepMeta, RowMetaInterface prev, String input[], String output[],
+    RowMetaInterface info, VariableSpace space, Repository repository, 
+    IMetaStore metaStore ) {
+    CheckResult cr;
+    if ( prev == null || prev.size() == 0 ) {
+      cr = new CheckResult( CheckResultInterface.TYPE_RESULT_WARNING, BaseMessages.getString( PKG, "${plugin_class_name}Meta.CheckResult.NotReceivingFields" ), stepMeta ); 
+      remarks.add( cr );
+    }
+    else {
+      cr = new CheckResult( CheckResultInterface.TYPE_RESULT_OK, BaseMessages.getString( PKG, "${plugin_class_name}Meta.CheckResult.StepRecevingData", prev.size() + "" ), stepMeta );  
+      remarks.add( cr );
+    }
+    
+    // See if we have input streams leading to this step!
+    if ( input.length > 0 ) {
+      cr = new CheckResult( CheckResultInterface.TYPE_RESULT_OK, BaseMessages.getString( PKG, "${plugin_class_name}Meta.CheckResult.StepRecevingData2" ), stepMeta ); 
+      remarks.add( cr );
+    }
+    else {
+      cr = new CheckResult( CheckResultInterface.TYPE_RESULT_ERROR, BaseMessages.getString( PKG, "${plugin_class_name}Meta.CheckResult.NoInputReceivedFromOtherSteps" ), stepMeta ); 
+      remarks.add( cr );
+    }
+  }
+  
+  public StepInterface getStep( StepMeta stepMeta, StepDataInterface stepDataInterface, int cnr, TransMeta tr, Trans trans ) {
+    return new ${plugin_class_name}( stepMeta, stepDataInterface, cnr, tr, trans );
+  }
+  
+  public StepDataInterface getStepData() {
+    return new ${plugin_class_name}Data();
+  }
+
+  public String getDialogClassName() {
+    return "${package}.ui.trans.steps.${rootArtifactId}.${plugin_class_name}Dialog";
+  }
+}
