+++ pentaho-pdi-step-plugin-archetype/src/main/resources/archetype-resources/impl/src/main/resources/__packageInPathFormat__/trans/steps/__rootArtifactId__/messages/messages_en_US.properties
@@ -1,0 +1,21 @@
+#set( $symbol_pound = '#' )
+#set( $symbol_dollar = '$' )
+#set( $symbol_escape = '\' )
+#####################################################################
+##
+##  ${plugin_class_name}
+##
+#####################################################################
+${plugin_class_name}.Name=${plugin_name} 
+${plugin_class_name}.Description=${plugin_description} 
+${plugin_class_name}.Log.LineNumber=Linenr 
+
+#####################################################################
+##
+##  ${plugin_class_name}Meta
+##
+#####################################################################
+${plugin_class_name}Meta.CheckResult.NotReceivingFields=Not receiving any fields from previous steps\!
+${plugin_class_name}Meta.CheckResult.StepRecevingData=Step is connected to previous one, receiving {0} fields
+${plugin_class_name}Meta.CheckResult.StepRecevingData2=Step is receiving info from other steps.
+${plugin_class_name}Meta.CheckResult.NoInputReceivedFromOtherSteps=No input received from other steps\!
