--- pentaho-pdi-step-plugin-archetype/src/main/resources/archetype-resources/src/main/resources/messages/messages_en_US.properties
@@ -1,43 +1,0 @@
-#set( $symbol_pound = '#' )
-#set( $symbol_dollar = '$' )
-#set( $symbol_escape = '\' )
-#####################################################################
-##
-##  ${plugin_class_name}Dialog
-##
-#####################################################################
-${plugin_class_name}Dialog.Stepname.Label=Step Name\:
-${plugin_class_name}Dialog.Shell.Title=My Step
-${plugin_class_name}Dialog.ActionButton=Action
-${plugin_class_name}Dialog.GroupText=Fieldset
-${plugin_class_name}Dialog.TextField350=Textfield (350px)\:
-${plugin_class_name}Dialog.TextField250=Textfield (250px)\:
-${plugin_class_name}Dialog.TextField75=Textfield (75px)\:
-${plugin_class_name}Dialog.Dropdown=Dropdown\:
-${plugin_class_name}Dialog.Tab1=Tab 1
-${plugin_class_name}Dialog.Tab2=Tab 2
-${plugin_class_name}Dialog.Button=Button
-${plugin_class_name}Dialog.RadioButton1=Radio button 1
-${plugin_class_name}Dialog.RadioButton2=Radio button 2
-${plugin_class_name}Dialog.Checkbox1=Checkbox 1
-${plugin_class_name}Dialog.Checkbox2=Checkbox 2
-${plugin_class_name}Dialog.TableHeader=Column
-${plugin_class_name}Dialog.ComboBoxOption1=Option 1
-${plugin_class_name}Dialog.ComboBoxOption2=Option 2
-
-#####################################################################
-##
-##  ${plugin_class_name}
-##
-#####################################################################
-${plugin_class_name}.Log.LineNumber=Linenr 
-
-#####################################################################
-##
-##  ${plugin_class_name}Meta
-##
-#####################################################################
-${plugin_class_name}Meta.CheckResult.NotReceivingFields=Not receiving any fields from previous steps\!
-${plugin_class_name}Meta.CheckResult.StepRecevingData=Step is connected to previous one, receiving {0} fields
-${plugin_class_name}Meta.CheckResult.StepRecevingData2=Step is receiving info from other steps.
-${plugin_class_name}Meta.CheckResult.NoInputReceivedFromOtherSteps=No input received from other steps\!
